refactor(allocation): move A26.3 routes from (app) to (terminal)
Relocates the Strategic IPS allocation pages to match the terminal
unification plan — (app) is being converted to read-only.

Route changes:
- (app)/portfolio/profiles/              → (terminal)/allocation/
- (app)/portfolio/profiles/[p]/allocation → (terminal)/allocation/[profile]/

Overwrites the previous (terminal)/allocation/+page.svelte (pre-A26.2
tactical weights editor — broken by migration 0155 that dropped
min_weight/max_weight columns; superseded by propose→approve flow).

Link updates (frontend routes only; backend API paths unchanged):
- PortfolioListPanel sidebar link: /portfolio/profiles → /allocation
- Profile list cards: /portfolio/profiles/{p}/allocation → /allocation/{p}
- Detail breadcrumb: /portfolio/profiles → /allocation
- E2E test target URL updated

Visual validation:
- /allocation renders with NETZ/TERMINAL brutalist layout, 3 profile cards
- /allocation/moderate renders KPI row, 18-row Strategic table, Pending
  Proposal panel with diff bars + metrics (E[r] 12.79%, CVaR -7.50%,
  Sharpe 3.57, Feasible badge)
- Terminal top-nav ALLOC tab active
- Zero console errors beyond pre-existing favicon 404
