+++ frontends/wealth/src/routes/(terminal)/allocation/[profile]/+page.server.ts
